processMoniter/requirements.txt
# Backend (Django + DRF)
Django==5.2.5
djangorestframework==3.16.1

# Agent (optional for server; required to run Windows agent)
psutil==7.0.0
requests==2.32.5

# Optional: build Windows EXE for agent
# pyinstaller==6.11.1 